+++ lib/ui/esrimap/esri_map_models/esrimap_ai_search_model.dart
