+++ lib/ui/esrimap/esri_map_models/esrimap_search_category.dart
