+++ lib/ui/esrimap/esri_map_models/map_search_result.dart
