+++ lib/ui/esrimap/esri_map_services/esrimap_auth_handler.dart
