+++ lib/ui/esrimap/esri_map_services/esrimap_route_service.dart
@@ -5,7 +5,7 @@
 /// ============================================================================
 
 import 'package:arcgis_maps/arcgis_maps.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/models/map_search_result.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/esri_map_models/map_search_result.dart';
 import 'package:flutter/material.dart';
 
 /// Result data structure holding the outcome of a solved route calculation.
