+++ lib/ui/esrimap/esri_map_widgets/esrimap_scene.dart
